server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "cachetools>=6.1.0",
    "cryptography>=45.0.6",
    "fastapi>=0.116.1",
    "fastmcp>=2.11.2",
    "google-api-python-client>=1.8.4",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=0.5.3",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "ruff>=0.12.8",
    "tomlkit>=0.13.3",
    "motor>=3.6.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "langchain-core>=0.3.0",
    "groq>=0.11.0",
]
